organize in different files and change names
