+++ DashAI/front/src/styles/components/InputStyles.jsx
@@ -1,0 +1,6 @@
+import { TextField } from "@mui/material";
+import { styled } from "@mui/material/styles";
+
+export const Input = styled(TextField)(({ theme }) => ({
+  minWidth: "20vw",
+}));
